requirements.txt
langchain
streamlit
tiktoken
openai
pypdf
python_dotenv
pinecone-client
sentence_transformers
unstructured
streamlit-ace
google-search-results
streamlit-camera-input-live
codeinterpreterapi[all]
pyrebase4
firebase-admin
st-pages
streamlit_extras
asyncio
